refactor(ui): extract thread message-data helpers into chrome/thread/
First extraction of the thread split (shelf-core plan Task 1): the pure
message-part helpers (risk/approval maps, tool naming, run collapsing,
turn text, approval preview) move to chrome/thread/message-data.ts.
Behavior-preserving; VendoThread re-imports them.
