+++ packages/ui/src/chrome/thread/message-data.ts
@@ -1,0 +1,146 @@
+import type { ApprovalRequest, RiskLabel } from "@vendoai/core";
+import { isToolUIPart, type UIMessage } from "ai";
+import { previewArgs } from "../humanize.js";
+import { LONG_TEXT_CAP, truncateHead } from "../truncate.js";
+
+export function partData(part: UIMessage["parts"][number]): unknown {
+  return "data" in part ? part.data : part;
+}
+
+// ENG-216 — a stable placeholder for the in-thread synthesized ApprovalRequest's
+// required `createdAt`. The wire approval part carries no timestamp; this value
+// is never displayed (the card hides the context byline in-thread) and a fixed
+// constant replaces the former per-render `new Date()` that churned on every
+// re-render and broke deterministic tests.
+export const SYNTHESIZED_CREATED_AT = "1970-01-01T00:00:00.000Z";
+
+export function riskByCall(messages: UIMessage[]): Map<string, RiskLabel> {
+  const risks = new Map<string, RiskLabel>();
+  for (const message of messages) {
+    for (const part of message.parts) {
+      if (part.type !== "data-vendo-approval") continue;
+      const data = partData(part) as { toolCallId?: unknown; risk?: unknown };
+      if (typeof data.toolCallId === "string" && ["read", "write", "destructive"].includes(String(data.risk))) {
+        risks.set(data.toolCallId, data.risk as RiskLabel);
+      }
+    }
+  }
+  return risks;
+}
+
+/** Guard approval metadata by tool call — carried in the data-vendo-approval
+    part beside the native ai-SDK approval (whose own id is transport-local). */
+export function approvalByCall(messages: UIMessage[]): Map<string, {
+  approvalId?: string;
+  invalidatedGrant?: ApprovalRequest["invalidatedGrant"];
+}> {
+  const approvals = new Map<string, {
+    approvalId?: string;
+    invalidatedGrant?: ApprovalRequest["invalidatedGrant"];
+  }>();
+  for (const message of messages) {
+    for (const part of message.parts) {
+      if (part.type !== "data-vendo-approval") continue;
+      const data = partData(part) as {
+        toolCallId?: unknown;
+        approvalId?: unknown;
+        invalidatedGrant?: { id?: unknown; grantedAt?: unknown };
+      };
+      if (typeof data.toolCallId !== "string") continue;
+      approvals.set(data.toolCallId, {
+        ...(typeof data.approvalId === "string" ? { approvalId: data.approvalId } : {}),
+        ...(typeof data.invalidatedGrant?.id === "string"
+          && typeof data.invalidatedGrant.grantedAt === "string"
+          ? { invalidatedGrant: data.invalidatedGrant as NonNullable<ApprovalRequest["invalidatedGrant"]> }
+          : {}),
+      });
+    }
+  }
+  return approvals;
+}
+
+export function toolName(part: Extract<UIMessage["parts"][number], { toolCallId: string }>): string {
+  return part.type === "dynamic-tool" && "toolName" in part ? part.toolName : part.type.replace(/^tool-/, "");
+}
+
+/** The app-boundary title: the payload's `name`, else its first heading Text node. */
+export function appTitle(payload: unknown): string | undefined {
+  const named = (payload as { name?: unknown }).name;
+  if (typeof named === "string" && named.trim()) return named;
+  const nodes = (payload as { nodes?: Array<{ component?: string; props?: Record<string, unknown> }> }).nodes;
+  if (!Array.isArray(nodes)) return undefined;
+  for (const node of nodes) {
+    if (node.component === "Text" && node.props?.variant === "heading" && typeof node.props.text === "string") {
+      return node.props.text;
+    }
+  }
+  return undefined;
+}
+
+/** A stable signature for a tool part — same tool + same input = the same call. */
+function toolSignature(part: Extract<UIMessage["parts"][number], { toolCallId: string }>): string {
+  const input = "input" in part ? part.input : undefined;
+  let serialized: string;
+  try {
+    serialized = JSON.stringify(input);
+  } catch {
+    serialized = String(input);
+  }
+  return `${toolName(part)}::${serialized}`;
+}
+
+/** ENG-216 — collapse runs of consecutive identical tool chips (e.g. eight
+    `host_listClientDocuments` calls) into one entry carrying a count. The
+    latest part in the run is kept so the chip icon reflects the final state. */
+export function collapseToolRuns(
+  parts: UIMessage["parts"],
+): { part: UIMessage["parts"][number]; index: number; count: number }[] {
+  const items: { part: UIMessage["parts"][number]; index: number; count: number }[] = [];
+  parts.forEach((part, index) => {
+    const previous = items.at(-1);
+    if (
+      isToolUIPart(part)
+      && previous !== undefined
+      && isToolUIPart(previous.part)
+      && toolSignature(previous.part) === toolSignature(part)
+    ) {
+      previous.count += 1;
+      previous.part = part;
+      return;
+    }
+    items.push({ part, index, count: 1 });
+  });
+  return items;
+}
+
+/** The plain text a user turn carried, joined across its text parts — the seed
+    for "edit last message" (ENG-215). */
+export function userText(message: UIMessage): string {
+  return message.parts
+    .filter((part): part is Extract<UIMessage["parts"][number], { type: "text" }> => part.type === "text")
+    .map(part => part.text)
+    .join("");
+}
+
+/** What "copy this turn" yields for an assistant message: its text parts (the
+    markdown source), blank-line separated — tool beats and views don't copy. */
+export function assistantText(message: UIMessage): string {
+  return message.parts
+    .filter((part): part is Extract<UIMessage["parts"][number], { type: "text" }> => part.type === "text")
+    .map(part => part.text)
+    .join("\n\n");
+}
+
+/** ENG-216 — the in-thread approval preview is built client-side (the wire part
+    carries no descriptor), so format args as readable `Label: value` lines
+    instead of the raw JSON with literal \n escapes end users were reading. */
+export function preview(input: unknown): string {
+  // ENG-216 — readable `Label: value` lines instead of raw JSON. ENG-218 — then
+  // bound the result before it reaches the DOM: a huge argument blob (dumped
+  // rows, base64) otherwise renders unbounded inside the approval card's <pre>,
+  // blowing up layout and the node count.
+  const formatted = previewArgs(input);
+  return formatted.length > LONG_TEXT_CAP
+    ? `${truncateHead(formatted)}\n… (${(formatted.length / 1000).toFixed(0)}k chars, truncated)`
+    : formatted;
+}
